requirements.txt
fastapi
uvicorn
pyodbc
openai
python-dotenv
psycopg2-binary